Fix: 모임 통장 전체 진행률 퍼센트 계산 수정

+++ waffle/javabackend/src/main/java/com/d109/waffle/api/teamaccount/dto/TeamAccountListDto.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class TeamAccountListDto {
     private int id;
     private String name;
-    private int percent;
+    private double percent;
     private LocalDate endDay;
     private Boolean master;
